OBUDPST-56: Interface Stats
This change addresses the need to always handle interface
byte counters as 64-bit values, regardless of the machine
architecture. Some compiler warnings were cleaned up and
the server will now display a message, when verbose is
enabled, if an invalid setup request is received.

+++ udpst.c
@@ -313,7 +313,7 @@ int main(int argc, char **argv) {
         }
         for (i = 0; i < conf.maxConnections; i++)
                 init_conn(i, FALSE);
-        for (i = 0; i < MAX_JPAYLOAD_SIZE / sizeof(int); i++)
+        for (i = 0; i < (int) (MAX_JPAYLOAD_SIZE / sizeof(int)); i++)
                 ((int *) repo.randData)[i] = random();
 
         //
